[0.017]: [+:>>] Re-coded parts, thus to remove any useless reusable code.
1) CCEXP now have the getTableIndexByName() method, which converts Table
names to Table IDs. Any overloaded method that uses Table name or ID,
now convert the Table name to Table ID and uses always the Table ID
implementation. In this way duplicate code reuse is removed. This also
helped to correct a bug in a Table-ID based method.

2) I re-ordered the basic classes, by using just class declarations at the
start of .hpp file.