[0.018]: [+:>>] 1) CCEXP::checkDuplicatedNames(), 2) Fixing Status checking, 3) DemoTest simplification
1) Added the CCEXP::checkDuplicatedNames() in order to check if a Table
with a specific name exist twice in a CCEXP object.

2) Fixing Status checking in different methods. If an error is critical,
then the status does not update correct, and CCEXP object is returning
"status error" in any other sequential call.

3) At UnitTests Demo t00_Basics.cpp, I updated the Error checking thus
results to be a bit more understandable. This is due to the fact that
the function CCEXP::GetErrors() is now not setting the Status to READY
state (as it should be from the start). If a critical error occurs, then
the CCEXP object takes a wrong status (ACTIVE) and can't be used anymore;
this is done thus the user to be "forced" to check and correct the error,
as probably it would be something that he would not like.

* If the user want to continue after a critical error, a new function
called CCEXP::DBG_SetState(CCEXP& obj, 1) can be used to set the State
to READY(=1) after the error. This is done in t00_Basics.cpp when the
#define ERROR_STOP definition is commented. In this case, any error
that otherwise should stop the main execution, is overpassed, thus the
whole error-tests procedure to be completed.