[0.050]: [:+>>] API Redesign: Return Values now corresponding to what the client needs.
This was a major update on CCEXP. Till now API functions and CCEXP methods where
using "int" return type which was corresponding to errors. For client this was
a bit hard to use when info or data were requesting. I.e. in order to get the
Rows of an array, client should create a variable and pass it via reference to
the API function:
	size_t Rows; CCEXP::Rows(OBJ,"NAME", Rows);

The above syntax is a bit confusing. I think this should be written as:
	size_t Rows = CCEXP::Rows(OBJ,"NAME");

For this reason I updated all API functions (functions that are defined in the
.cpp file, as far as and the relative inline templates functions in the .hpp
file). Internal CCEXP methods however need no change (at least at this point).

As far as it concerns API functions that are returning non-standard types, but
template types (template class<T>) I decided that they should return references
to types instead of types.
If a Table contains non-standard type (I have not test this scenario but in
future I would like to support this) then returning a type would consist of a
copy/assignment of data. With references I can avoid this overhead. However
there are two problems that arise:

1) A bit more complex syntax. Instead of types, use references. I.e.
	Don't use: vector<float>  vf = CCEXP::getRow<float>(OBJ,"NAME",row);.
	Use      : vector<float>& vf = CCEXP::getRow<float>(OBJ,"NAME",row);
	vf is a reference, which is references directly the the CCEXP object row.
	Also now the template type of the Table is always required!

2) In case of errors, a returning reference can not be null. Instead of using
external libaries (i.e. boost::optional) I decided to create a new return object
with shared_ptr() that will be returned on error. If no error occur this object
will be destroyed automatically as it won't be used (shared_ptr does this job).
If error occur the client will get a reference to that new empty object. When he
stops using it then the object will be destroyed. It's up to the user to check
for errors using the __CCEXP_ERR_DISPLAY macro (or a relative implementation
like this). Client can check if there was an error using the added CCEXP
object variable named ErrorId. Thus the OBJ.ErrorId should be always 0 after a
call to API functions. If not, then the returned data probably are not correct
and should not be used. Notice that CCEXP.ErrorId is always set to 0 after an API
function call.